xbuild/cargo: Use LLVM `libc++` std library on all platforms
Even though we pre-package and provide a sysroot for Android, iOS and
MacOS - which are all based on LLVM's `libc++` stdlib, the compiler
still tries to use the one from `gcc-c++` by default and fails otherwise
for not even finding the standard headers (that are, again, provided in
the prepackaged sysroot).

Telling the compiler to use this library through `-stdlib` resolves
those header inclusion errors, without having to install `gcc-c++`.