Append `runtime_libs` specified in manifest to APK
Certain apps and crates `dlopen` libraries at runtime (which are not
listed in the shared library table by design, hence not found by
`llvm-readobj`).  Facilitate this by allowing the user to specify a list
of folders that need to be scanned for libraries to include in the final
package at while building.  This folder structure includes the platform
ABI to select the desired shared libraries.