adb: Pass `-W` (wait for app start) to `am start` before getting `pidof` (#118)
Sometimes the app hasn't started yet, and `xbuild` fails with:

    Error: failed to get pid:

because `pidof` could not find any process with that name, and exits
with a non-zero error code (and stderr is left empty).  Simply passing
`-W` to make `am start` block a little while longer until the app is
actually running seems to resolve this issue in the common case.