gradle: Embed assets in `baseAssets` "Play Asset Delivery" pack
Assets for AABs could be embedded in the application pack but have
significant limitations and size constraints.  Implement the most
simple form of an asset pack that comprises all assets listed in
`manifest.yaml` in an `install-time` pack that is a fixed dependency of
the app.  This makes the files available to `AAssetManager` as soon as
the app is started while being exempt of the 150MB limit.

This matches the existing "asset embedding" functionality provided by
our native APK build.

https://developer.android.com/guide/playcore/asset-delivery/integrate-native