Fix missing `"exe"` variant and copy-paste mistakes in `Format`
Passing `--format exe` would fail, complaining that the "arch" `exe`
is not supported.  Copy-paste mistakes from "arch" to "format" aside,
the `"exe"` string wasn't in the match arm.

Realistically, as we are already utilizing `clap`, we should remove
all this error-prone (proven by this example) open-coding of broken
parsers and utilize their derives to automatically generate conversion
functions.

This should at the same time assist us in generating help files, as the
hardcoded documentation for `--format` currently states that `exe` is a
supported value.