Android: Bump default target SDK version to `34`
The Google Play Store now requires apps to target at least SDK level 34
to be compliant with the latest "limitations and requirements".  Set our
default just as high so that apps can be published "out of the box".

This does imply that all apps built with `xbuild`'s default
target SDK level are supposed to be aware of the latest behaviour
changes that they opt into when specifying `targetSdkVersion=34`:
https://developer.android.com/about/versions/14/behavior-changes-14
Fortunately, almost none of these changes appear to affect the minimal
API surface that is reachable by the NDK.  Any app using them is likely
written in Kotlin or Java and managed/compiled via `gradle`, without
involving `xbuild` at all.