template: Update to `dioxus 0.4`/`wry 0.28` with `libwebkit2gtk-4.1-dev`
We've been getting deprecation warnings in the CI that Ubuntu 22.04
is going to be replaced with Ubuntu 24.04 soon for the `ubuntu-latest`
image.

That happened just a few days ago, and now the CI complains
that `libwebkit2gtk-4.0-dev` can no longer be found in
Ubuntu Noble's APT repositories.  We have to upgrade to
`4.1`, which can be achieved by using at least `wry 0.25`:
https://github.com/tauri-apps/wry/commit/c5f3b36

This is used by `dioxus 0.4`, which is the target of this upgrade.  As I
am not a user of any of the `wry`/`dioxus` functionality inside `xbuild`
(and if not for me, it doesn't look like anyone else is maintaining or
caring about this either), I kept this upgrade as small as possible and
didn't upgrade to `dioxus 0.6.1` (the latest as of writing) which might
incur more breaking changes and require testing which I'm not inclined
to perform.  All I really need is the CI to succeed again :)