Use `gradle` when the build target is `Aab` (#156)
Instead of explicitly requiring `gradle: true` in the manifest, enable
it by default when the (implicit or explicit!) output package format
is `Aab` for convenience, as there is currently no way to select the
`gradle` backend via command line parameters and hardcoding it in
`manifest.yaml` makes it inconvenient to build `Apk`s with the "native"
builtin backend.

Also insert validation in case the user explicitly inserts `gradle:
false` but tries to build `Aab`s.  The default for releases is still
`Aab` if there is an explicit `gradle: true`, but an `Apk` is built
otherwise.