apk/res: Parse new sparse `Type` chunk since Android 34
Android 34 replaced most `Type` chunks with few entries with a new
`SPARSE` list.  Here the offset table is now a combination of a `u16`
array index and a `u16` byte offset, allowing it to take up way less
space in binary form when few elements are set while still reaching
high(er) indexes (typically one or a handful of elements with an index
in the hundreds or thousands).