apk/res: Handle `Type` and `TypeSpec` headers in the parser and writer
The header for `TypeSpec` and `Type`, as well as the "index" mapping
contain byte offsets, reserved fields and flags that are only usable
for the binary parser and writer.  Even though handling these are
cumbersome at writing time (because of jumping back to the header or
offsets array to write accurate byte positions after first serializing
prior elements), that's still a lot more correct than leaving the user
to **guess** what byte offsets to put into the vector/headers.

That is otherwise going to be even more more complicated with the
introduction of a new `SPARSE` `Type` flag in Android 34, and `OFFSET16`
`Type` flag in Android 35.