apk/res: Take offset and `entries_start` into account when parsing `Type` chunks
The original code was assuming that all non-empty `Type` entries were
stored sequentially, directly after the index list and tightly packed
together with no padding in between.  Previously added assertions
support this case, but this will no longer be fully correct with the
introduction of a sparse `Type` in Android 34, and 16-bit offsets in
Android 35.

Note that a future patch should be removing the unusable index array
that is exposed to the user: this byte-offset is no longer relevant
to them when the elements have been parsed, and is terribly hard and
error-prone to set up manually before serializing.  This however becomes
tedious to deal with in our writer.